Add new centralized types folder
Managing them was becoming a nightmare...
