--- client/src/routes/auth/types.tsx
@@ -1,10 +1,0 @@
-export interface UserInfo {
-  email: string;
-  password: string;
-}
-
-export interface LoginInfo extends UserInfo {
-  rememberMe: boolean;
-}
-
-export type RegisterInfo = UserInfo;
